% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmap_drug.R
\docType{data}
\name{cmap_drug}
\alias{cmap_drug}
\title{the CMap drug MesH ID}
\format{
A data frame with 1129 rows and 4 variables:
\describe{
\item{cmap_name}{drug names, this is the drug orignal names}
\item{CID}{drug pubchem ID}
\item{meshterm}{drug MeSH name, parsed by pyMeSHSim}
\item{meshID}{drug MeSH ID, the drug MeSH ID}
...
}
}
\usage{
cmap_drug
}
\description{
A dataset containing the cmap drug name and MeSH term ID
there are 1309 drugs in CMap, 1129 of them can be assigned MeSH ID
cmap_drug
}
\keyword{datasets}
