% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDisease2drugDb.R
\name{disease2drugDb}
\alias{disease2drugDb}
\title{This is for load disease2drug.db}
\usage{
disease2drugDb(data_type)
}
\arguments{
\item{data_type}{get disease2drug table or drug2gene table, the value is "Disease" or "Drug"}
}
\description{
This is for load disease2drug.db
}
