% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease2pathway.R
\name{disease2pathway}
\alias{disease2pathway}
\title{Enriched the pathways for diseases}
\usage{
disease2pathway(DiseaseMeSHID)
}
\arguments{
\item{DiseaseMeSHID}{the MeSH id of intersted disease}
}
\value{
dataframe it containing the pathways enriched
}
\description{
Enriched the pathways for diseases
}
\examples{
disease2pathway("D015179")


}
