% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug2gene.R
\name{drug2gene}
\alias{drug2gene}
\title{Select the drug related genes}
\usage{
drug2gene(DrugMeSHID)
}
\arguments{
\item{DrugMeSHID}{the MeSH id of intersted drug}
}
\value{
dataframe
}
\description{
Select the drug related genes
}
\examples{
drug2gene("D016861")


}
