% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_gene_df.R
\docType{data}
\name{pathway_gene_df}
\alias{pathway_gene_df}
\title{the genes included in each pathway}
\format{
A data frame with 506182 rows and 2 variables:
\describe{
\item{pathway_name}{pathway names, MSigDB C5 GO biological process}
\item{gene}{one gene in the pathway, gene symbol}
...
}
}
\source{
\url{https://www.gsea-msigdb.org/gsea/msigdb/}
}
\usage{
pathway_gene_df
}
\description{
A dataset containing the genes and pathways
pathway_gene_df
}
\keyword{datasets}
