% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease2drug.R
\name{disease2drug}
\alias{disease2drug}
\title{Select the disease related drugs}
\usage{
disease2drug(DiseaseMeSHID)
}
\arguments{
\item{DiseaseMeSHID}{the MeSH id of intersted disease, only Main heading diseases were allowed}
}
\value{
dataframe
}
\description{
Select the disease related drugs
}
\examples{
disease2drug("D015179")


}
