% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease2drug.R
\name{disease2drug}
\alias{disease2drug}
\title{Select the disease related drugs}
\usage{
disease2drug(DiseaseMeSHID, Repository = "DSEATM")
}
\arguments{
\item{DiseaseMeSHID}{The MeSH id of intersted disease, only Main heading diseases are allowed here.}

\item{Repository}{one of the values in c("ALL", "DSEATM").
"ALL" means all MeSH terms in "D" category will be obtained in the disease-drug association.
In this condition, all the terms are chemical and drugs.
"DSEATM" means only MeSH terms below "ChemicalActionsandUses", "OrganicChemicals" and "Compounds,Heterocyclic"

The default is "DSEATM"}
}
\value{
dataframe It contained all drugs associated to a specific diseases.
}
\description{
Select the disease related drugs
}
\examples{
#here, we only give out the drug ID in the table,
#detail information for each drug can be retrived by pyMeSHSim
#get disease-drug association in DSEATM repository
df = disease2drug("D015179")
# when using ALL repository,
# the disease-drug associations are in indeed disease-chemical associations
# Users can filter the drugs by their own principles in this data set.
df = disease2drug("D015179", Repository="ALL")

#get CMap drugs associated to a disease
df = disease2drug("D015179", Repository="ALL")
df_cmap = df[df$Drug \%in\% DSEATM::cmap_drug$meshID,]



}
