% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease2gene.R
\name{disease2gene}
\alias{disease2gene}
\title{Select the disease related genes by drugs}
\usage{
disease2gene(DiseaseMeSHID)
}
\arguments{
\item{DiseaseMeSHID}{the MeSH id of intersted disease}
}
\value{
vactor it containing Entrez gene IDs related to the diseases.
}
\description{
Select the disease related genes by drugs
}
\examples{
disease2gene("D015179")


}
